[Refactor]#2 Storage 클래스들을 data패키지로 이동
 - 추후 local data source로 변경될 예정
