+++ app/src/main/java/com/gallery/kakaogallery/data/DataStorage.kt
@@ -1,4 +1,4 @@
-package com.gallery.kakaogallery.util
+package com.gallery.kakaogallery.data
 
 import com.gallery.kakaogallery.presentation.application.KakaoGalleryApplication
