+++ app/src/main/java/com/gallery/kakaogallery/presentation/application/KakaoGalleryApplication.kt
@@ -1,7 +1,7 @@
 package com.gallery.kakaogallery.presentation.application
 
 import android.app.Application
-import com.gallery.kakaogallery.util.DataStorage
+import com.gallery.kakaogallery.data.DataStorage
 import com.gallery.kakaogallery.presentation.network.NetworkUtil
 import com.gallery.kakaogallery.presentation.network.RetrofitManager
